Move exceptions to their own namespace
