+++ utils/migrator/src/Exceptions/FileIOException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace UniEngine\Utils\Migrations;
+namespace UniEngine\Utils\Migrations\Exceptions;
 
 class FileIOException extends \Exception {}
