+++ utils/migrator/src/Exceptions/FileMissingException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace UniEngine\Utils\Migrations;
+namespace UniEngine\Utils\Migrations\Exceptions;
 
 class FileMissingException extends \Exception {}
